refactor(api): move reverse-geocoding into @op/common
Keep the taxonomy router slim: the Google Geocoding call and its cache wrapper
now live in a reverseGeocodeLocation service in @op/common (sibling of
resolveBoundary), and the router only declares its input/output schemas and
delegates. Matches how the resolveBoundary router is structured.

Addresses @scazan review note on PR #1262.
